// cms-kit-monorepo-to-extract-payload-project-from/apps/payload/src/payload-types.ts
export interface Page {
  id: number;
  _abPassPercentage?: number | null;
  /**
   * The title of the page
   */
  title: string;
  /**
   * The header to display on the page
   */
  header?: (number | null) | Header;
  /**
   * The footer to display on the page
   */
  footer?: (number | null) | Footer;
  blocks: (
    | HeroBlock
    | TextSectionBlock
    | ContentBlock
    | FaqBlock
    | TestimonialsListBlock
    | CardsGridBlock
    | CarouselBlock
    | LogosBlock
    | LinksListBlock
  )[];
  meta?: {
    title?: string | null;
    /**
     * Maximum upload file size: 12MB. Recommended file size for images is <500KB.
     */
    image?: (number | null) | Media;
    description?: string | null;
    /**
     * Allow search engines to index this page
     */
    robots?: ("index" | "noindex") | null;
  };
  /**
   * When enabled, the slug will auto-generate from the title field on save and autosave.
   */
  generateSlug?: boolean | null;
  slug: string;
  parent?: (number | null) | Page;
  breadcrumbs?:
    | {
        doc?: (number | null) | Page;
        url?: string | null;
        label?: string | null;
        id?: string | null;
      }[]
    | null;
  /**
   * The original page this variant belongs to.
   */
  _abVariantOf?: (number | null) | Page;
  _abVariantPercentages?:
    | {
        [k: string]: unknown;
      }
    | unknown[]
    | string
    | number
    | boolean
    | null;
  folder?: (number | null) | FolderInterface;
  updatedAt: string;
  createdAt: string;
  _status?: ("draft" | "published") | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "header".
 */
export interface Header {
  id: number;
  /**
   * The name of the header
   */
  name?: string | null;
  /**
   * The logo to display in the header
   */
  logo: number | Media;
  /**
   * Navigation items in the header (up to 6 items)
   */
  navItems?:
    | {
        /**
         * Link settings
         */
        link: {
          type?: ("reference" | "custom" | "customPage") | null;
          newTab?: boolean | null;
          reference?:
            | ({
                relationTo: "page";
                value: number | Page;
              } | null)
            | ({
                relationTo: "posts";
                value: number | Post;
              } | null);
          url?: string | null;
          customPage?: ("blog" | "search") | null;
          label: string;
        };
        id?: string | null;
      }[]
    | null;
  updatedAt: string;
  createdAt: string;
  _status?: ("draft" | "published") | null;
}
// ...
export interface Footer {
  id: number;
  /**
   * The name of the footer
   */
  name: string;
  /**
   * The logo to display in the footer
   */
  logo: number | Media;
  /**
   * Footer navigation links (up to 10 items)
   */
  links?:
    | {
        /**
         * Link settings
         */
        link: {
          type?: ("reference" | "custom" | "customPage") | null;
          newTab?: boolean | null;
          reference?:
            | ({
                relationTo: "page";
                value: number | Page;
              } | null)
            | ({
                relationTo: "posts";
                value: number | Post;
              } | null);
          url?: string | null;
          customPage?: ("blog" | "search") | null;
          label: string;
        };
        id?: string | null;
      }[]
    | null;
  /**
   * Footer body text
   */
  text?: {
    root: {
      type: string;
      children: {
        type: any;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ("ltr" | "rtl") | null;
      format: "left" | "start" | "center" | "right" | "end" | "justify" | "";
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  } | null;
  /**
   * Copyright text shown at the bottom
   */
  copywriteText?: string | null;
  updatedAt: string;
  createdAt: string;
  _status?: ("draft" | "published") | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "HeroBlock".
 */
export interface HeroBlock {
  title?: string | null;
  richText?: {
    root: {
      type: string;
      children: {
        type: any;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ("ltr" | "rtl") | null;
      format: "left" | "start" | "center" | "right" | "end" | "justify" | "";
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  } | null;
  actions?:
    | {
        type?: ("reference" | "custom" | "customPage") | null;
        newTab?: boolean | null;
        reference?:
          | ({
              relationTo: "page";
              value: number | Page;
            } | null)
          | ({
              relationTo: "posts";
              value: number | Post;
            } | null);
        url?: string | null;
        customPage?: ("blog" | "search") | null;
        label: string;
        /**
         * Choose how the link should be rendered.
         */
        appearance?: ("default" | "outline") | null;
        id?: string | null;
      }[]
    | null;
  image: {
    image: number | Media;
    aspectRatio?:
      | (
          | "16/9"
          | "3/2"
          | "4/3"
          | "1/1"
          | "9/16"
          | "1/2"
          | "4/1"
          | "3/1"
          | "auto"
        )
      | null;
  };
  enabled?: boolean | null;
  color?: ("black" | "white") | null;
  /**
   * Overlay opacity (0-100)
   */
  opacity?: number | null;
  section?: {
    theme?: ("light" | "dark" | "light-gray" | "dark-gray") | null;
    paddingY?: ("none" | "base" | "large") | null;
    paddingX?: ("none" | "base") | null;
    maxWidth?: ("none" | "base") | null;
    background?: {
      /**
       * Upload an image or video. Use the "Background" folder.
       */
      media?: (number | null) | Media;
      overlay?: ("black" | "white") | null;
      /**
       * 0 = transparent, 100 = fully opaque
       */
      opacity?: number | null;
    };
  };
  id?: string | null;
  blockName?: string | null;
  blockType: "hero";
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "TextSectionBlock".
 */
export interface TextSectionBlock {
  text: {
    root: {
      type: string;
      children: {
        type: any;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ("ltr" | "rtl") | null;
      format: "left" | "start" | "center" | "right" | "end" | "justify" | "";
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  };
  section?: {
    theme?: ("light" | "dark" | "light-gray" | "dark-gray") | null;
    paddingY?: ("none" | "base" | "large") | null;
    paddingX?: ("none" | "base") | null;
    maxWidth?: ("none" | "base") | null;
    background?: {
      /**
       * Upload an image or video. Use the "Background" folder.
       */
      media?: (number | null) | Media;
      overlay?: ("black" | "white") | null;
      /**
       * 0 = transparent, 100 = fully opaque
       */
      opacity?: number | null;
    };
  };
  id?: string | null;
  blockName?: string | null;
  blockType: "textSection";
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "ContentBlock".
 */
export interface ContentBlock {
  heading?: string | null;
  layout: "image-text" | "text-image";
  image: number | Media;
  content: {
    root: {
      type: string;
      children: {
        type: any;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ("ltr" | "rtl") | null;
      format: "left" | "start" | "center" | "right" | "end" | "justify" | "";
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  };
  section?: {
    theme?: ("light" | "dark" | "light-gray" | "dark-gray") | null;
    paddingY?: ("none" | "base" | "large") | null;
    paddingX?: ("none" | "base") | null;
    maxWidth?: ("none" | "base") | null;
    background?: {
      /**
       * Upload an image or video. Use the "Background" folder.
       */
      media?: (number | null) | Media;
      overlay?: ("black" | "white") | null;
      /**
       * 0 = transparent, 100 = fully opaque
       */
      opacity?: number | null;
    };
  };
  id?: string | null;
  blockName?: string | null;
  blockType: "content";
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "FaqBlock".
 */
export interface FaqBlock {
  heading: string;
  items: {
    question: string;
    answer: {
      root: {
        type: string;
        children: {
          type: any;
          version: number;
          [k: string]: unknown;
        }[];
        direction: ("ltr" | "rtl") | null;
        format: "left" | "start" | "center" | "right" | "end" | "justify" | "";
        indent: number;
        version: number;
      };
      [k: string]: unknown;
    };
    id?: string | null;
  }[];
  section?: {
    theme?: ("light" | "dark" | "light-gray" | "dark-gray") | null;
    paddingY?: ("none" | "base" | "large") | null;
    paddingX?: ("none" | "base") | null;
    maxWidth?: ("none" | "base") | null;
    background?: {
      /**
       * Upload an image or video. Use the "Background" folder.
       */
      media?: (number | null) | Media;
      overlay?: ("black" | "white") | null;
      /**
       * 0 = transparent, 100 = fully opaque
       */
      opacity?: number | null;
    };
  };
  id?: string | null;
  blockName?: string | null;
  blockType: "faq";
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "TestimonialsListBlock".
 */
export interface TestimonialsListBlock {
  heading?: string | null;
  subheading?: string | null;
  testimonialItems?:
    | {
        testimonial: number | Testimonial;
        id?: string | null;
      }[]
    | null;
  /**
   * The duration of the animation in seconds. Default is 60 seconds.
   */
  duration?: number | null;
  showRating?: boolean | null;
  showAvatar?: boolean | null;
  section?: {
    theme?: ("light" | "dark" | "light-gray" | "dark-gray") | null;
    paddingY?: ("none" | "base" | "large") | null;
    paddingX?: ("none" | "base") | null;
    maxWidth?: ("none" | "base") | null;
    background?: {
      /**
       * Upload an image or video. Use the "Background" folder.
       */
      media?: (number | null) | Media;
      overlay?: ("black" | "white") | null;
      /**
       * 0 = transparent, 100 = fully opaque
       */
      opacity?: number | null;
    };
  };
  id?: string | null;
  blockName?: string | null;
  blockType: "testimonialsList";
}
// ...
export interface CardsGridBlock {
  columns?: number | null;
  items: {
    title: string;
    description?: string | null;
    image?: {
      image?: (number | null) | Media;
      aspectRatio?:
        | (
            | "16/9"
            | "3/2"
            | "4/3"
            | "1/1"
            | "9/16"
            | "1/2"
            | "4/1"
            | "3/1"
            | "auto"
          )
        | null;
    };
    link?: {
      type?: ("reference" | "custom" | "customPage") | null;
      newTab?: boolean | null;
      reference?:
        | ({
            relationTo: "page";
            value: number | Page;
          } | null)
        | ({
            relationTo: "posts";
            value: number | Post;
          } | null);
      url?: string | null;
      customPage?: ("blog" | "search") | null;
      label?: string | null;
      /**
       * Choose how the link should be rendered.
       */
      appearance?: ("default" | "outline") | null;
    };
    alignVariant?: ("left" | "center" | "right") | null;
    rounded?: ("none" | "large") | null;
    backgroundColor?:
      | ("none" | "light" | "dark" | "light-gray" | "dark-gray" | "gradient-2")
      | null;
    id?: string | null;
  }[];
  section?: {
    theme?: ("light" | "dark" | "light-gray" | "dark-gray") | null;
    paddingY?: ("none" | "base" | "large") | null;
    paddingX?: ("none" | "base") | null;
    maxWidth?: ("none" | "base") | null;
    background?: {
      /**
       * Upload an image or video. Use the "Background" folder.
       */
      media?: (number | null) | Media;
      overlay?: ("black" | "white") | null;
      /**
       * 0 = transparent, 100 = fully opaque
       */
      opacity?: number | null;
    };
  };
  id?: string | null;
  blockName?: string | null;
  blockType: "cardsGrid";
}